@@ -1,13 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace ASPNETCoreMVCSecurity.Models
+﻿namespace ASPNETCoreMVCSecurity.Models
 {
-    public class Book
-    {
-        public string Title { get; set; }
-        public string Publisher { get; set; }
-    }
+    public record Book(string Title, string? Publisher);
+
 }
