@@ -1,3 +1,3 @@
-@using ASPNETCoreMVCSecurity
+﻿@using ASPNETCoreMVCSecurity
 @using ASPNETCoreMVCSecurity.Models
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
